added action_sign_test.go and merge action_verifysignature to action_sign
