+++ action_sign.go
@@ -1,5 +1,9 @@
 package holochain
 
+import (
+	ic "github.com/libp2p/go-libp2p-crypto"
+)
+
 //------------------------------------------------------------
 // Sign
 
@@ -22,5 +26,37 @@ func (a *APIFnSign) Call(h *Holochain) (response interface{}, err error) {
 		return
 	}
 	response = sig.B58String()
+	return
+}
+
+//------------------------------------------------------------
+// VerifySignature
+
+type APIFnVerifySignature struct {
+	b58signature string
+	data         string
+	b58pubKey    string
+}
+
+func (a *APIFnVerifySignature) Name() string {
+	return "verifySignature"
+}
+
+func (a *APIFnVerifySignature) Args() []Arg {
+	return []Arg{{Name: "signature", Type: StringArg}, {Name: "data", Type: StringArg}, {Name: "pubKey", Type: StringArg}}
+}
+
+func (a *APIFnVerifySignature) Call(h *Holochain) (response interface{}, err error) {
+	var b bool
+	var pubKey ic.PubKey
+	sig := SignatureFromB58String(a.b58signature)
+
+	pubKey, err = DecodePubKey(a.b58pubKey)
+
+	b, err = h.VerifySignature(sig, a.data, pubKey)
+	if err != nil {
+		return
+	}
+	response = b
 	return
 }
