+++ action_update_test.go
@@ -1,7 +1,6 @@
 package holochain
 
 import (
-	. "github.com/holochain/holochain-proto/hash"
 	peer "github.com/libp2p/go-libp2p-peer"
 	. "github.com/smartystreets/goconvey/convey"
 	"testing"
